+++ Source/Helpers.swift
@@ -43,4 +43,29 @@ func logInit(object: AnyObject) {
 /// Log the termination ("de-initialization" in Swift terms) of an object.
 func logTerm(object: AnyObject) {
     print("<TERM> \(unsafeAddressOf(object)) (\(object.dynamicType)) \(object.description)")
+
+// MARK: - Collection shuffling
+// Taken from <http://stackoverflow.com/questions/24026510/how-do-i-shuffle-an-array-in-swift>.
+
+extension CollectionType {
+    /// Return a copy of `self` with its elements shuffled.
+    func shuffle() -> [Generator.Element] {
+        var list = Array(self)
+        list.shuffleInPlace()
+        return list
+    }
+}
+
+extension MutableCollectionType where Index == Int {
+    /// Shuffle the elements of `self` in-place.
+    mutating func shuffleInPlace() {
+        // Empty and single-element collections don't shuffle.
+        if count < 2 { return }
+        
+        for i in 0..<count - 1 {
+            let j = Int(arc4random_uniform(UInt32(count - i))) + i
+            guard i != j else { continue }
+            swap(&self[i], &self[j])
+        }
+    }
 }